+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.35] - 2024-12-19
+
+### Changed
+- Person detail: Work history is now sorted by start date descending (most recent first)
+- Person detail: Current positions appear at the top of the work history list
+
 ## [1.0.34] - 2024-12-19
 
 ### Fixed